+++ src/org/opensha2/util/NshmpSiteFiles.java
@@ -30,6 +30,8 @@ import java.util.Set;
 import org.opensha2.calc.NamedLocation;
 import org.opensha2.geo.Location;
 import org.opensha2.geo.LocationList;
+import org.opensha2.util.GeoJson.Feature;
+import org.opensha2.util.GeoJson.FeatureCollection;
 
 import com.google.common.base.Function;
 import com.google.common.base.Functions;
@@ -42,11 +44,11 @@ import com.google.gson.GsonBuilder;
 import com.google.gson.annotations.SerializedName;
 
 /**
- * Methods for exporting data as TXT, JSON, KML, CSV, etc.
+ * Methods for exporting nshmp site data as CSV and GeoJSON.
  *
  * @author Peter Powers
  */
-final class Export {
+final class NshmpSiteFiles {
 
 	private static final Gson GSON = new GsonBuilder()
 		.setPrettyPrinting()
@@ -55,6 +57,9 @@ final class Export {
 
 	private static final Path EXPORT_DIR = Paths.get("etc", "nshm");
 
+	/**
+	 * Updates all site list and map files in etc/nshm.
+	 */
 	public static void main(String[] args) throws IOException {
 		writeNshmpSites();
 		writeNshmpPolys();
@@ -158,7 +163,7 @@ final class Export {
 		List<Feature> features = new ArrayList<>();
 		int i = 0;
 		for (LocationList coords : coordList) {
-			features.add(createPolygon(
+			features.add(GeoJson.createPolygon(
 				nameList.get(i++),
 				coords,
 				Optional.<String> absent(),
@@ -166,7 +171,7 @@ final class Export {
 		}
 		FeatureCollection fc = new FeatureCollection();
 		fc.features = features;
-		String json = cleanPoly(GSON.toJson(fc));
+		String json = GeoJson.cleanPoly(GSON.toJson(fc));
 		Files.write(out, json.getBytes(StandardCharsets.UTF_8));
 	}
 
@@ -179,20 +184,20 @@ final class Export {
 
 		List<Feature> features = new ArrayList<>();
 		if (bounds != null) {
-			features.add(createPolygon(
+			features.add(GeoJson.createPolygon(
 				name + " Map Extents",
 				bounds,
 				Optional.of(GeoJson.Value.EXTENTS),
 				Optional.<Double> absent()));
 		}
-		features.add(createPolygon(
+		features.add(GeoJson.createPolygon(
 			name,
 			coords,
 			Optional.<String> absent(),
 			Optional.of(spacing)));
 		FeatureCollection fc = new FeatureCollection();
 		fc.features = features;
-		String json = cleanPoly(GSON.toJson(fc));
+		String json = GeoJson.cleanPoly(GSON.toJson(fc));
 		Files.write(out, json.getBytes(StandardCharsets.UTF_8));
 	}
 
@@ -244,115 +249,29 @@ final class Export {
 
 		List<Feature> features = new ArrayList<>(sites.size());
 		for (NamedLocation loc : sites) {
-			features.add(createPoint(loc));
+			features.add(GeoJson.createPoint(loc));
 		}
 		FeatureCollection fc = new FeatureCollection();
 		fc.features = features;
-		String json = cleanPoints(GSON.toJson(fc));
+		String json = GeoJson.cleanPoints(GSON.toJson(fc));
 		Files.write(out, json.getBytes(StandardCharsets.UTF_8));
 	}
+	
+//	private static void writeJsonSites(Path out, Collection<? extends NamedLocation> sites)
+//			throws IOException {
+//
+//		List<Feature> features = new ArrayList<>(sites.size());
+//		for (NamedLocation loc : sites) {
+//			features.add(createPoint(loc));
+//		}
+//		FeatureCollection fc = new FeatureCollection();
+//		fc.features = features;
+//		String json = cleanPoints(GSON.toJson(fc));
+//		Files.write(out, json.getBytes(StandardCharsets.UTF_8));
+//	}
+
 
 	/* GeoJSON objects */
 
-	static class FeatureCollection {
-		String type = "FeatureCollection";
-		List<Feature> features;
-	}
-
-	static class Feature {
-		String type = "Feature";
-		String id;
-		Geometry geometry = new Geometry();
-		Properties properties;
-	}
-
-	static Feature createPoint(NamedLocation loc) {
-		Feature f = new Feature();
-		f.geometry.type = "Point";
-		f.geometry.coordinates = toCoordinates(loc.location());
-		f.properties = new PointProperties();
-		f.properties.title = loc.toString();
-		return f;
-	}
-
-	private static final String EXTENTS_COLOR = "#4169E1";
-	
-	static Feature createPolygon(
-			String name,
-			LocationList coords,
-			Optional<String> id,
-			Optional<Double> spacing) {
-		
-		Feature f = new Feature();
-		
-		PolyProperties properties = new PolyProperties();
-		properties.title = name;
-		if (spacing.isPresent()) {
-			properties.spacing = spacing.get();
-		}
-		
-		if (id.isPresent()) {
-			f.id = id.get();
-			coords = coords.bounds().toList();
-			properties.strokeColor = EXTENTS_COLOR;
-			properties.fillColor = EXTENTS_COLOR;
-		}
-		
-		f.geometry.type = "Polygon";
-		f.geometry.coordinates = ImmutableList.of(toCoordinates(coords));
-		f.properties = properties;
-		return f;
-	}
-
-	static class Geometry {
-		String type;
-		Object coordinates;
-	}
-
-	static class Properties {
-		String title;
-	}
-
-	static class PointProperties extends Properties {
-		@SerializedName("marker-size")
-		String markerSize = "small";
-	}
-
-	static class PolyProperties extends Properties {
-		Double spacing;
-		@SerializedName("fill")
-		String fillColor;
-		@SerializedName("marker-color")
-		String strokeColor;
-	}
-
-	static double[] toCoordinates(Location loc) {
-		return new double[] { round(loc.lon(), 5), round(loc.lat(), 5) };
-	}
-
-	static double[][] toCoordinates(LocationList locs) {
-		double[][] coords = new double[locs.size()][2];
-		for (int i = 0; i < locs.size(); i++) {
-			coords[i] = toCoordinates(locs.get(i));
-		}
-		return coords;
-	}
-
-	/* brute force compaction of coordinate array onto single line */
-	static String cleanPoints(String s) {
-		return s.replace(": [\n          ", ": [")
-			.replace(",\n          ", ", ")
-			.replace("\n        ]", "]") + "\n";
-	}
-
-	/* brute force compaction of coordinate array onto single line */
-	static String cleanPoly(String s) {
-		return s
-			.replace("\n          [", "[")
-			.replace("[\n              ", "[ ")
-			.replace(",\n              ", ", ")
-			.replace("\n            ]", " ]")
-			.replace("\n        ]", "]") + "\n";
-	}
 
 }
